backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2.4
django-admin==2.0.2
django-allauth==65.10.0
django-cors-headers==4.7.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
excel-base==1.0.4
idna==3.10
isoweek==1.3.3
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.4
screen==1.0.1
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
tzlocal==5.3.1
urllib3==2.5.0
xlwt==1.3.0
